refactor: use reverse in tests and move tests to folders
